Moved layer commands into canvas crate.
